+++ digitalmall/digitalmall-portal/src/main/java/com/thezjh/digitalmallportal/entity/OrdersItem.java
@@ -9,18 +9,18 @@ import lombok.*;
 @Data
 public class OrdersItem {
 
-  private long id;
-  private long orderId;
-  private long userId;
-  private long productId;
-  private String productName;
-  private String productImage;
-  private double currentPrice;
-  private long quantity;
-  private double price;
-  private java.sql.Timestamp createTime;
-  private java.sql.Timestamp updateTime;
-  private long addressId;
+    private Integer id;
+    private long orderId;
+    private long userId;
+    private long productId;
+    private String productName;
+    private String productImage;
+    private double currentPrice;
+    private long quantity;
+    private double price;
+    private java.sql.Timestamp createTime;
+    private java.sql.Timestamp updateTime;
+    private long addressId;
 
 
 }
