+++ digitalmall/digitalmall-portal/src/main/java/com/thezjh/digitalmallportal/entity/Products.java
@@ -9,16 +9,16 @@ import lombok.*;
 @Data
 public class Products {
 
-  private long id;
-  private String productName;
-  private long status;
-  private String title;
-  private String image;
-  private double price;
-  private String detail;
-  private long stock;
-  private java.sql.Timestamp createTime;
-  private java.sql.Timestamp updateTime;
+    private Integer id;
+    private String productName;
+    private long status;
+    private String title;
+    private String image;
+    private double price;
+    private String detail;
+    private long stock;
+    private java.sql.Timestamp createTime;
+    private java.sql.Timestamp updateTime;
 
 
 }
